frontend/requirements.txt
streamlit>=1.30.0
requests>=2.31.0
pandas>=2.1.0
plotly>=5.18.0
python-dotenv>=1.0.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.23.0
streamlit-autorefresh>=1.0.1
pyyaml>=6.0.1
